Patched the loader for newer GCC on Ubuntu 19.10.
It seems like the linker now splits sections in pieces that may be smaller than a page, causing the
loader to try to load the same page multiple times and fail. Now, offsets are handled properly, and
loading of the new kind of files work properly. If the first part of the page is read only, and the
second is read-write, then the entire page will not be read-write, which could cause issues. At the
moment, it seems this only happens for the text section, which is read only anyway, so it should not
be a big issue in the near future at least.

Signed-off-by: Filip Strömbäck <filip.stromback@liu.se>