Fixed compilation issue with newer GCC (currently on Arch)
Newer GCC seems to include an extra section when building the loader (.note.gnu....) with a fairly
large offset. This caused the loader.bin file to be way more than 512 bytes, which in turn caused
the build process to make an incorrect disk image, resulting in boot failure.

We now build the loader with objdump so that we can remove the extra sections easier (we could
probably do it with flags to ld, but I could not find them). So now, it works even on later GCC versions.

Signed-off-by: Filip Strömbäck <filip.stromback@liu.se>